+++ NEWS.md
@@ -1,3 +1,7 @@
+# CHANGES IN VERSION 1.7.2
+   - Add Seqinfo to Imports
+   - Move GenomeInfoDb from Imports to Suggests
+
 # CHANGES IN VERSION 1.7.1
    - Updated test function
    - Bump up version number to trigger Bioconductor build